+++ test/examples/tests_latex.jl
@@ -2,7 +2,7 @@ using Test
 
 # DOCUMENTER_TEST_EXAMPLES can be used to control which builds are performed in
 # make.jl, and we need to set it to the relevant LaTeX builds.
-ENV["DOCUMENTER_TEST_EXAMPLES"] = "latex_simple"
+ENV["DOCUMENTER_TEST_EXAMPLES"] = "latex latex_simple"
 
 # When the file is run separately we need to include make.jl which actually builds
 # the docs and defines a few modules that are referred to in the docs. The make.jl
@@ -22,6 +22,14 @@ end
         @test isa(doc, Documenter.Documents.Document)
         let build_dir = joinpath(examples_root, "builds", "latex_simple")
             @test joinpath(build_dir, "DocumenterLaTeXSimple.pdf") |> isfile
+        end
+    end
+
+    @testset "PDF/LaTeX" begin
+        doc = Main.examples_latex_doc
+        @test isa(doc, Documenter.Documents.Document)
+        let build_dir = joinpath(examples_root, "builds", "latex")
+            @test joinpath(build_dir, "DocumenterLaTeX.pdf") |> isfile
         end
     end
 end